Fixed bug in touch scaling on device with no Full-HD resolution and on-screen button. Networking 3 It's official: I hate with all my heart "play games library"

+++ app/src/main/java/com/example/bizzi/GameSystem/GameObSubSystem/DrawableComponent.java
@@ -48,10 +48,10 @@ public class DrawableComponent extends Component {
     }
     
     private void buildBorder(){
-        DEST.left = (int) (x-semiWidth);
-        DEST.top = (int) (y-semiHeight);
-        DEST.right = (int) (x + semiWidth) ;
-        DEST.bottom = (int) (y + semiHeight);
+        DEST.left = x-semiWidth;
+        DEST.top = y-semiHeight;
+        DEST.right = x + semiWidth;
+        DEST.bottom = y + semiHeight;
         
     }
